support workloads across multiple hosts using same /26
This commit tries to make calico-bgp-daemon works identical with BIRD backend.

When multiple hosts are sharing same /26, the prefix is only advertised
from the host which used the prefix first.

To get reachability, succeeding hosts needs to advertise /32 routes
which are taken from the /26 and assigned to each workloads.
The /32 routes can be learned by subscribing the netlink updates.

The first host which used the prefix needs to suppress advertising /32
routes and only advertise the /26 route.
Export policy('calico_aggr') is used to implement this behavior.

calico_aggr consist of two statements.
The first statement allows the /26 routes to be advertised, and the
second one rejects the /32 routes.

you can check the current policy by using gobgp command on calico/node

$ gobgp policy
Name calico_aggr:
    StatementName calico_aggr_stmt0:
      Conditions:
        PrefixSet: any aggregated
      Actions:
         accept
    StatementName calico_aggr_stmt1:
      Conditions:
        PrefixSet: any host
      Actions:
         reject

$ gobgp policy prefix
NAME        PREFIX
aggregated  192.168.1.0/26 26..26
host        192.168.1.0/26 26..32

Signed-off-by: ISHIDA Wataru <ishida.wataru@lab.ntt.co.jp>